requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
rembg==2.0.56
Pillow==10.2.0